Add testing framework + test behavior of generateResourcesFile
