+++ package.json
@@ -27,10 +27,13 @@
     "mkdirp": "^1.0.3",
     "replace-in-file": "^4.0.0",
     "rimraf": "^2.6.3",
-    "standard": "^14.3.3"
+    "standard": "^14.3.3",
+    "tap": "^14.10.7",
+    "tmp": "^0.2.1"
   },
   "scripts": {
     "clean": "rimraf build/",
+    "test": "tap test/*.js",
     "data-files-ad-block-rust": "node scripts/generateAdBlockRustDataFiles.js",
     "data-files-ad-block": "npm run --prefix ./node_modules/ad-block data-files",
     "data-files-https-everywhere": "npm run --prefix ./node_modules/https-everywhere-builder build",
